+++ src/features/achs/achs.vue
@@ -37,8 +37,8 @@ import { ACH_NAMES, achs, hasAch, getAchCount } from "./achs";
 import { formatWhole } from "@/util/format";
 import Tooltip from "@/components/Tooltip.vue";
 
-const ACH_ROWS = 4;
-const ACH_COLS = 7;
+const ACH_ROWS = 5;
+const ACH_COLS = 8;
 </script>
 
 <style scoped>