+++ src/flourish/tabs/tabs.vue
@@ -178,4 +178,14 @@ import { availableTabs, setTab } from "./tabs";
   background-color: hsl(300, 50%, 35%);
   border-color: hsl(300, 50%, 60%);
 }
+
+.btn.collapse {
+  background-color: hsl(267, 51%, 20%);
+  border-color: hsl(244, 50%, 40%);
+}
+
+.btn.collapse:hover {
+  background-color: hsl(235, 50%, 35%);
+  border-color: hsl(254, 50%, 60%);
+}
 </style>